inference/requirements.txt
FastAPI
TensorFlow
Pillow
numpy
uvicorn
aiofiles